move getChangedFiles to new file
